refactor(extract components): extract StatusTag component
